pyproject.toml
[project]
name = "llmgine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42.0",
    "google-cloud>=0.34.0",
    "google-cloud-aiplatform>=1.75.0",
    "loguru>=0.7.3",
    "notebook>=7.3.2",
    "notion-client>=2.3.0",
    "openai>=1.59.3",
    "pydantic>=2.10.4",
    "pyperclip>=1.9.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.6",
    "streamlit>=1.41.1",
    "textual>=1.0.0",
    "textual-dev>=1.7.0",
    "typer>=0.15.1",
    "setuptools>=75.6.0",
    "prompt-toolkit>=3.0.50",
    "pygments>=2.19.1",
    "contextvars>=2.4",
    "memory-profiler>=0.61.0",
]

[project.scripts]
llmgine = "programs.router:main"

[dependency-groups]
dev = [
    "llmgine",
]

[tool.uv.sources]
llmgine = { workspace = true }
